+++ GUI/WindowFrames/AchievementsFrame/AchievementsFrameMixin.lua
@@ -1,46 +1,5 @@
 -- [[ Namespaces ]] --
 local _, addon = ...;
-addon.GUI.AchievementsFrame = {};
-local achievementsFrame = addon.GUI.AchievementsFrame;
-
-function achievementsFrame:Load()
-	local frame = CreateFrame("Frame", "KrowiAF_AchievementsFrame", AchievementFrame, "KrowiAF_AchievementsFrame_Template");
-	frame:SetPoint("TOPLEFT", AchievementFrameCategories, "TOPRIGHT", 22, 0);
-	frame:SetPoint("BOTTOM", 0, 20);
-	frame:SetPoint("RIGHT", -20, 0);
-
-	tinsert(ACHIEVEMENTFRAME_SUBFRAMES, frame:GetName());
-
-	local scrollFrame = frame.ScrollFrame;
-	local scrollBar = scrollFrame.ScrollBar;
-	local scrollBarShow = getmetatable(scrollBar).__index.Show;
-	scrollBar.Show = function()
-		frame.Show_Hide(frame, scrollBarShow, -46);
-	end;
-	local scrollBarHide = getmetatable(scrollBar).__index.Hide;
-	scrollBar.Hide = function()
-		frame.Show_Hide(frame, scrollBarHide, -20);
-	end;
-
-	scrollFrame.update = function()
-		frame:Update();
-	end
-
-	local template = "KrowiAF_AchievementButton_" .. (addon.Options.db.Achievements.Compact and "Small" or "Normal") .. "_Template";
-	HybridScrollFrame_CreateButtons(scrollFrame, template, 0, -2);
-	local buttons = scrollFrame.buttons;
-    for _, button in next, buttons do
-		button:PostLoad(scrollFrame);
-	end
-
-	hooksecurefunc("AchievementFrameAchievements_ForceUpdate", function()
-		frame:ForceUpdate();
-	end); -- Issue #3: Fix
-
-	frame.ScrollBarStep = scrollBar:GetValueStep();
-
-	addon.GUI.AchievementsFrame = frame; -- Overwrite with the actual frame since all functions are injected to it
-end
 
 KrowiAF_AchievementsFrameMixin = {};
 
@@ -122,7 +81,7 @@ function KrowiAF_AchievementsFrameMixin:Update()
 		end
 	end
 
-	local buttonCollapsedHeight = addon.Options.db.Achievements.ButtonCollapsedHeight;
+	local buttonCollapsedHeight = buttons[1].CollapsedHeight;
 	local totalHeight = #cachedAchievements * buttonCollapsedHeight;
 	local extraHeight = scrollFrame.largeButtonHeight or buttonCollapsedHeight;
 	totalHeight = totalHeight + extraHeight - buttonCollapsedHeight;
@@ -151,7 +110,7 @@ function KrowiAF_AchievementsFrameMixin.ClearHighlightedButton()
 end
 
 function KrowiAF_AchievementsFrameMixin:ExpandSelection(button)
-	HybridScrollFrame_ExpandButton(self.ScrollFrame, ((button.index - 1) * addon.Options.db.Achievements.ButtonCollapsedHeight), button:GetHeight());
+	HybridScrollFrame_ExpandButton(self.ScrollFrame, ((button.index - 1) * button.CollapsedHeight), button:GetHeight());
 	self:Update();
 	self:AdjustSelection();
 end