+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Protection so Blizzard Statistics can not be in front of Guild, would result in errors
 - When a criteria updates for an achievement or you earn an achievement, this will also visually update without having to manually refresh the window
 - Keybind to open the current zone
+- Right click menu to the summary achievements and calendar side window achievements
 
 ### Changed
 - Tabs are now handled by addon and tab name and API is accessible by other addons (see Plugins/Overachiever.lua for example or ask for assistance on discord)